update plugin installation

+++ fuel/app/tasks/installplugins.php
@@ -22,26 +22,35 @@ class InstallPlugins
      */
     public function createCss()
     {
+        $message = \Cli::color('Looking for CSS to install', 'green');
+        \Cli::write($message);
+
         $p = new \Mustached\Plugins;
         $csss = $p->getCss();
 
-        $message = \Cli::color('Checkin for CSS to install ...', 'green');
-        \Cli::write($message);
-
-        foreach ($csss as $plugin => $css) {
-            $localDir = DOCROOT.'public/assets/css/plugins/'.$plugin.DIRECTORY_SEPARATOR;
-            if (!file_exists($localDir)) {
-                \File::create_dir(DOCROOT.'public/assets/css/plugins/', $plugin);
+        if (!empty($csss)) {
+            foreach ($csss as $plugin => $css) {
+                $localDir = DOCROOT.'public/assets/css/plugins/'.$plugin.DIRECTORY_SEPARATOR;
+                if (!file_exists($localDir)) {
+                    \File::create_dir(DOCROOT.'public/assets/css/plugins/', $plugin);
+                }
+                if (!file_exists($localDir.$css['version'].'.css')) {
+                    \File::copy(APPPATH.DIRECTORY_SEPARATOR.'modules'.DIRECTORY_SEPARATOR.$plugin.DIRECTORY_SEPARATOR.$css['path'], $localDir.$css['version'].'.css');
+                        $message = \Cli::color('Plugin '.$plugin.': CSS v'.$css['version'].' created', 'green');
+                        \Cli::write($message);
+                }    
             }
-            if (!file_exists($localDir.$css['version'].'.css')) {
-                \File::copy(APPPATH.DIRECTORY_SEPARATOR.'modules'.DIRECTORY_SEPARATOR.$plugin.DIRECTORY_SEPARATOR.$css['path'], $localDir.$css['version'].'.css');
-                    $message = \Cli::color('Plugin '.$plugin.': CSS v'.$css['version'].' created', 'green');
-                    \Cli::write($message);
-            }    
+
+            $message = \Cli::color('CSS installed', 'green');
+            \Cli::write($message);            
+        } else {
+            $message = \Cli::color('No CSS found in the plugins', 'green');
+            \Cli::write($message);            
         }
 
-        $message = \Cli::color('... CSS installation finished!', 'green');
-        \Cli::write($message);
+        
+
+
     }
 
     /**
@@ -54,23 +63,26 @@ class InstallPlugins
         $p = new \Mustached\Plugins;
         $plugins = $p->getPlugins();        
 
-        $message = \Cli::color('Checkin for images to install ...', 'green');
+        $message = \Cli::color('Looking for images to install', 'green');
         \Cli::write($message);
 
-        foreach ($plugins as $plugin) {
-            $pluginDir = APPPATH.'modules/'.$plugin.'/assets/img';
-            $localDir  = DOCROOT.'public/assets/img/plugins/'.$plugin;
-            if (file_exists($pluginDir)) {
-                if (file_exists($localDir)) {
-                    \File::delete_dir($localDir, true);    
-                }                
-                \File::copy_dir($pluginDir, $localDir);
-                $message = \Cli::color('Plugin '.$plugin.': images created', 'green');
-                \Cli::write($message);
-            }            
+        if (!empty($plugins)) {
+            foreach ($plugins as $plugin) {
+                $pluginDir = APPPATH.'modules/'.$plugin.'/assets/img';
+                $localDir  = DOCROOT.'public/assets/img/plugins/'.$plugin;
+                if (file_exists($pluginDir)) {
+                    if (file_exists($localDir)) {
+                        \File::delete_dir($localDir, true);    
+                    }                
+                    \File::copy_dir($pluginDir, $localDir);
+                    $message = \Cli::color('Plugin '.$plugin.': images created', 'green');
+                    \Cli::write($message);
+                }            
+            }
+            $message = \Cli::color('Images installation finished', 'green');
+            \Cli::write($message);
         }
-
-        $message = \Cli::color('... Images installation finished!', 'green');
+        $message = \Cli::color('No images found in the plugins', 'green');
         \Cli::write($message);
     }    
 }
